simple_requirements.txt
scikit-learn
numpy
pandas
Pillow
matplotlib
seaborn
joblib